+++ src/shared/Auth/WowCrypt.h
@@ -23,8 +23,8 @@ public:
 	WowCrypt();
 	~WowCrypt();
 
-	const static int CRYPTED_SEND_LEN = 4;
-	const static int CRYPTED_RECV_LEN = 6;
+	const static size_t CRYPTED_SEND_LEN = 4;
+	const static size_t CRYPTED_RECV_LEN = 6;
 
 	void Init();
